Add better examples for CustomSnackbar and UndoSnackbar
